refactor(tests): move the empty batch allocation request
